Refactor utils imports to be consistent
