+++ server/chat-commands/room-settings.ts
@@ -6,7 +6,7 @@
  *
  * @license MIT
  */
-import * as Utils from '../../lib/utils';
+import {Utils} from '../../lib';
 import type {EffectiveGroupSymbol, RoomPermission} from '../user-groups';
 
 const RANKS = Config.groupsranking;
