+++ server/chat-plugins/calculator.ts
@@ -1,4 +1,4 @@
-import * as Utils from '../../lib/utils';
+import {Utils} from '../../lib';
 
 type Operator = '^' | '%' | '/' | '*' | '+' | '-';
 interface Operators {
