+++ server/chat-plugins/jeopardy.ts
@@ -1,4 +1,4 @@
-import * as Utils from '../../lib/utils';
+import {Utils} from '../../lib';
 
 const BACKGROUND_COLOR = "#0000FF";
 const HEIGHT = 40;
