+++ server/chat-plugins/othermetas.ts
@@ -5,7 +5,7 @@
  * @author Kris
 */
 
-import * as Utils from '../../lib/utils';
+import {Utils} from '../../lib';
 
 interface StoneDeltas {
 	baseStats: {[stat in StatName]: number};
