+++ server/chat-plugins/poll.ts
@@ -2,7 +2,7 @@
  * Poll chat plugin
  * By bumbadadabum and Zarel.
  */
-import * as Utils from '../../lib/utils';
+import {Utils} from '../../lib';
 
 const MINUTES = 60000;
