+++ server/chat-plugins/room-events.ts
@@ -6,7 +6,7 @@
  *
  * @license MIT license
  */
-import * as Utils from '../../lib/utils';
+import {Utils} from '../../lib';
 
 export interface RoomEvent {
 	eventName: string;
