+++ server/chat-plugins/uno.ts
@@ -6,7 +6,7 @@
  *
  * @license MIT license
  */
-import * as Utils from '../../lib/utils';
+import {Utils} from '../../lib';
 
 type Color = 'Green' | 'Yellow' | 'Red' | 'Blue' | 'Black';
 interface Card {
